refactor(graphql): Drop test-only reset and redundant subscriber bookkeeping
Split option-matrix tests into process-isolated files instead of resetting
and re-subscribing the process-global channels.
